chrisa/libusdt#5: free memory associated with provider build
We were assuming that typical use cases wouldn't care if we leaked a bit
of memory, but that wasn't a safe assumption: dynamic creation and removal
of providers requires us to run cleanly.

This change frees memory allocated for the provider and probes, but does
require that consumers call the new usdt_probe_release() routine on each
usdt_probedef_t when they're done with it.